Balance/license icon with simplified label.
