+++ app/lib/frontend/templates/views/pkg/package_list.dart
@@ -16,6 +16,8 @@ import '../../../static_files.dart' show staticUrls;
 import '../../_consts.dart';
 import '../../package_misc.dart';
 
+import 'license.dart';
+
 /// Renders the listing page (list of packages).
 d.Node listOfPackagesNode({
   required SearchForm? searchForm,
@@ -82,7 +84,7 @@ d.Node _packageItem(
     ];
   }
 
-  final licenseNode = _license(view.spdxIdentifiers);
+  final licenseNode = packageListMetadataLicense(view.spdxIdentifiers);
   final releases = view.releases!;
   final metadataNode = d.fragment([
     d.span(
@@ -146,31 +148,6 @@ d.Node _packageItem(
             ))
         .toList(),
   );
-}
-
-/// Nicely formatted labels for common SPDX identifiers.
-const _spdxLabels = {
-  'AGPL-3.0': 'AGPL 3.0',
-  'Apache-2.0': 'Apache 2.0',
-  'BSD-2-Clause': 'BSD 2-clause',
-  'BSD-3-Clause': 'BSD 3-clause',
-  'GPL-3.0': 'GPL 3.0',
-  'MPL-2.0': 'MPL 2.0',
-  'LGPL-2.1': 'LGPL 2.1',
-  'LGPL-3.0': 'LGPL 3.0',
-};
-String _spdxLabel(String id) => _spdxLabels[id] ?? id;
-
-d.Node? _license(List<String>? spdxIdentifiers) {
-  if (spdxIdentifiers == null || spdxIdentifiers.isEmpty) {
-    return null;
-  }
-  if (spdxIdentifiers.length == 1) {
-    return d.text('${_spdxLabel(spdxIdentifiers.single)} license');
-  } else {
-    final more = spdxIdentifiers.length - 1;
-    return d.text('${_spdxLabel(spdxIdentifiers.first)} +$more licenses');
-  }
 }
 
 d.Node _item({
