+++ app/test/frontend/golden/pkg_admin_page.html
@@ -442,7 +442,7 @@
             </p>
             <h3 class="title">License</h3>
             <p>
-              BSD-3-Clause (
+              BSD 3-clause (
               <a href="/packages/oxygen/license">LICENSE</a>
               )
             </p>
@@ -499,7 +499,7 @@
           </p>
           <h3 class="title">License</h3>
           <p>
-            BSD-3-Clause (
+            BSD 3-clause (
             <a href="/packages/oxygen/license">LICENSE</a>
             )
           </p>
