+++ app/test/frontend/golden/pkg_changelog_page.html
@@ -266,7 +266,7 @@
             </p>
             <h3 class="title">License</h3>
             <p>
-              BSD-3-Clause (
+              BSD 3-clause (
               <a href="/packages/oxygen/license">LICENSE</a>
               )
             </p>
@@ -324,7 +324,7 @@
           </p>
           <h3 class="title">License</h3>
           <p>
-            BSD-3-Clause (
+            BSD 3-clause (
             <a href="/packages/oxygen/license">LICENSE</a>
             )
           </p>
