+++ app/test/frontend/golden/pkg_example_page.html
@@ -261,7 +261,7 @@
             </p>
             <h3 class="title">License</h3>
             <p>
-              BSD-3-Clause (
+              BSD 3-clause (
               <a href="/packages/oxygen/license">LICENSE</a>
               )
             </p>
@@ -319,7 +319,7 @@
           </p>
           <h3 class="title">License</h3>
           <p>
-            BSD-3-Clause (
+            BSD 3-clause (
             <a href="/packages/oxygen/license">LICENSE</a>
             )
           </p>
