+++ app/test/frontend/golden/pkg_score_page.html
@@ -353,7 +353,7 @@
             </p>
             <h3 class="title">License</h3>
             <p>
-              BSD-3-Clause (
+              BSD 3-clause (
               <a href="/packages/oxygen/license">LICENSE</a>
               )
             </p>
@@ -411,7 +411,7 @@
           </p>
           <h3 class="title">License</h3>
           <p>
-            BSD-3-Clause (
+            BSD 3-clause (
             <a href="/packages/oxygen/license">LICENSE</a>
             )
           </p>
