+++ app/test/frontend/golden/pkg_show_page_flutter_plugin.html
@@ -248,7 +248,7 @@
             </p>
             <h3 class="title">License</h3>
             <p>
-              BSD-3-Clause (
+              BSD 3-clause (
               <a href="/packages/flutter_titanium/license">LICENSE</a>
               )
             </p>
@@ -310,7 +310,7 @@
           </p>
           <h3 class="title">License</h3>
           <p>
-            BSD-3-Clause (
+            BSD 3-clause (
             <a href="/packages/flutter_titanium/license">LICENSE</a>
             )
           </p>
