+++ app/test/frontend/golden/pkg_versions_page.html
@@ -367,7 +367,7 @@
             </p>
             <h3 class="title">License</h3>
             <p>
-              BSD-3-Clause (
+              BSD 3-clause (
               <a href="/packages/oxygen/license">LICENSE</a>
               )
             </p>
@@ -425,7 +425,7 @@
           </p>
           <h3 class="title">License</h3>
           <p>
-            BSD-3-Clause (
+            BSD 3-clause (
             <a href="/packages/oxygen/license">LICENSE</a>
             )
           </p>
