+++ app/test/frontend/golden/publisher_packages_page.html
@@ -228,7 +228,10 @@
                           <img class="package-vp-icon" src="/static/hash-%%etag%%/img/verified-publisher-icon.svg" alt="shield icon for verified publishers" width="14" height="14" title="Published by a pub.dev verified publisher"/>
                           <a href="/publishers/example.com">example.com</a>
                         </span>
-                        <span class="packages-metadata-block">unknown license</span>
+                        <span class="packages-metadata-block">
+                          <img class="packages-metadata-license-icon" src="/static/img/material-icon-balance-48.svg" alt="Icon for licenses." width="16" height="16"/>
+                          unknown
+                        </span>
                       </p>
                       <div>
                         <div class="-pub-tag-badge">
@@ -290,7 +293,10 @@
                           <span class="-x-ago" title="%%neon-created-date%%">%%x-ago%%</span>
                           )
                         </span>
-                        <span class="packages-metadata-block">BSD 3-clause license</span>
+                        <span class="packages-metadata-block">
+                          <img class="packages-metadata-license-icon" src="/static/img/material-icon-balance-48.svg" alt="Icon for licenses." width="16" height="16"/>
+                          BSD 3-clause
+                        </span>
                       </p>
                       <div>
                         <div class="-pub-tag-badge">
