+++ app/test/frontend/golden/search_page.html
@@ -402,7 +402,10 @@
                   <span class="-x-ago" title="%%oxygen-created-date%%">%%x-ago%%</span>
                   )
                 </span>
-                <span class="packages-metadata-block">BSD 3-clause license</span>
+                <span class="packages-metadata-block">
+                  <img class="packages-metadata-license-icon" src="/static/img/material-icon-balance-48.svg" alt="Icon for licenses." width="16" height="16"/>
+                  BSD 3-clause
+                </span>
               </p>
               <div>
                 <div class="-pub-tag-badge">
@@ -474,7 +477,10 @@
                   <span class="-x-ago" title="%%oxygen-created-date%%">%%x-ago%%</span>
                   )
                 </span>
-                <span class="packages-metadata-block">BSD 3-clause license</span>
+                <span class="packages-metadata-block">
+                  <img class="packages-metadata-license-icon" src="/static/img/material-icon-balance-48.svg" alt="Icon for licenses." width="16" height="16"/>
+                  BSD 3-clause
+                </span>
               </p>
               <div>
                 <div class="-pub-tag-badge">
