+++ pkg/web_css/lib/src/_list.scss
@@ -134,6 +134,15 @@
     margin: 0 8px 0 0;
   }
 
+  .packages-metadata-license-icon {
+    display: inline-block;
+    height: 14px;
+    position: relative;
+    top: 3px;
+    opacity: 0.75;
+    margin-right: 3px;
+  }
+
   .packages-api {
     color: #3c4043;
     font-size: 12px;
